LPS-94184 Move tests to Journal
