+++ modules/apps/fragment/fragment-entry-processor/fragment-entry-processor-editable-test/build.gradle
@@ -3,12 +3,8 @@ dependencies {
 	testIntegrationCompile group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
 	testIntegrationCompile group: "org.jsoup", name: "jsoup", version: "1.10.2"
 	testIntegrationCompile project (":apps:asset:asset-api")
-	testIntegrationCompile project (":apps:dynamic-data-mapping:dynamic-data-mapping-api")
-	testIntegrationCompile project (":apps:dynamic-data-mapping:dynamic-data-mapping-test-util")
 	testIntegrationCompile project (":apps:fragment:fragment-api")
 	testIntegrationCompile project (":apps:fragment:fragment-service")
-	testIntegrationCompile project (":apps:journal:journal-api")
-	testIntegrationCompile project (":apps:journal:journal-test-util")
 	testIntegrationCompile project(":core:petra:petra-string")
 	testIntegrationCompile project(":test:arquillian-extension-junit-bridge")
 }
