+++ modules/apps/journal/journal-test/src/testIntegration/resources/com/liferay/journal/dependencies/fragment_entry_link_mapped_ddm.json
